Add supersearchfacet command (#2)
This adds a supersearchfacet command that lets you access facet
information via the Supersearch API. It supports a few output formats.

It also has a "--daily" mode that runs the facet for each day in the
given date range.

This is a little clunky still, but it seems like it works.

Issue #2