Add support for python 3.10, drop 3.6, fix dev requirements
This adds support for Python 3.10 by adding it to the tox.ini.

This drops support for Python 3.6.

This fixes CI to run a matrix strategy so it's doing the tests per
Python version in parallel.

This fixes dev requirements so they're defined in `setup.py` rather than
a separate `requirements-dev.txt` file.