Add support for a remainder value in facets
Sometimes we have facet values, but then we also want to know the number
of crash reports that don't have that field at all. This isn't normally
handled by facets, but we can derive it from the search count minus all
the value counts.

This adds a remainder value count denoted by "--" to supersearchfacet.
This is very helpful for flags.