Fix non-string sorting
When I added the "--" thing (which is incredibly helpful), I didn't take
into account that not all values are strings. When the values aren't
strings (for example, application_build_id values), then it would kick
up an error when trying to sort things.

This adds an AlwaysFirst class which we use to smooth over sorting "--"
value with other things. It also guarantees it always shows up first in
the list.