Add supersearch tests (#16)
This adds tests for the supersearch command.

This changes the default back to "format=tab" instead of "format=table".

This sets "_facets_size=0" so it doesn't get back facet information
which it ignores.

This fixes "format=table" to show/hide headers.

This fixes "format=tab" to use click.echo which doesn't do fancy-pants
stuff.