Rework output for supersearch and supersearchfacet (#39)
This reworks the output for supersearch and supersearchfacet so that
format=tab can stream as results are coming in. This is helpful for
incremental progress when using it with pipes.

While doing this, I redid parts of supersearchfacet so that it only
supports a single _facet and if none are specified, it uses the default
that Super Search does which is signature.

I also fixed json output to spit out the records including the -- which
covers remaining count.

I also fixed tox to pass {posargs} to pytest making it easier to debug
specific test cases.