Redo output for fetch-data (#106)
This reworks the output for fetch-data so each line clearly specifies
which crash id it's related to. Further, it removes a lot of the
coloring. It also removes start/end indicators. The end result is a lot
less output where all the important information is in the line and it
can be read left-to-right and sorted and make sense.

While doing this, I adjusted fetch-data to use "crash_id" instead of
both "crashid" and "crash_id".